requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
websockets
python-multipart
mlx-whisper
